Cleanup cmd directory
The cmd directory in a go project by convention is a place to put all of
the compile targets for the project. If we're going to use the cmd
directory, we should also follow this convention.

This PR:
- moves the main module for riff to `cmd/riff`. The file is symlined
  back to the project root to preserve backwards compatibility with
  `go get`
- moves the commands directory from `cmd/commands`, to
  `pkg/commands/riff`
- updates all imports that need updating
