+++ pkg/riff/commands/completion.go
