+++ pkg/riff/commands/function.go
