+++ pkg/riff/commands/namespace_test.go
@@ -8,9 +8,9 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
-	"github.com/projectriff/riff/cmd/commands"
 	"github.com/projectriff/riff/pkg/core"
 	"github.com/projectriff/riff/pkg/core/mocks"
+	"github.com/projectriff/riff/pkg/riff/commands"
 	"github.com/spf13/cobra"
 )
