+++ pkg/riff/commands/table_display.go
